requirements.txt
aioboto3==8.2.0
aiobotocore==1.1.2
aiofiles==0.5.0
boto3==1.14.44
botocore==1.17.44
fastapi==0.62.0
python-dateutil==2.8.1
python-decouple==3.3
uvicorn==0.11.8
websockets==8.1
